+++ src/components/_common/Booth.js
@@ -4,7 +4,7 @@ import { styled } from "styled-components";
 import { useNavigate } from "react-router-dom";
 
 //부스 대표 사진
-import boothCover from "../../assets/images/mypage/boothcover.png";
+import boothCover from "../../assets/images/Mypage/boothcover.png";
 import { ReactComponent as PinkHeart } from "../../assets/icons/heart-empty.svg";
 import Degree from "../_common/Degree";
