+++ src/components/_common/Concert.js
@@ -3,7 +3,7 @@ import "../../App.css";
 import { styled } from "styled-components";
 
 //부스 대표 사진
-import concert from "../../assets/images/mypage/concert.png";
+import concert from "../../assets/images/Mypage/concert.png";
 import { ReactComponent as PinkHeart } from "../../assets/icons/heart-empty.svg";
 import { useNavigate } from "react-router-dom";
